Message order fix
The order of the messages in the active chat were starting from newest
to oldest.
This meant when the `ActiveChat` component rendered the messages, the
newest message was at the top.

FIX:
Reconfigured the query to order the messages by their createdAt time in
the ascending direction.

OTHER:
When including the Message model on line 24, it cannot handle the order
option.
Therefore I removed it, as it's confusing to the reader.