New messages get added to convo
The reason for why new messages for pre-existing conversations were not
getting updated ui state was because of how the state was being changed
in `addMessageToConversation` in the Home component.

React uses `Object.is()` to determine whether the state object has
changed.
The issue is that the code was not creating a shallow clone of the
converations state, which meant that the hook didn't re-render the Home
component.

FIX:
Setting the state to a shallow clone by using `conversations.map()` and
updating the appropriate conversation with an updated shallow copy