Feature: Read status - SocketIO event impl,API fix
Added a SocketIO event called `read-conversation`. It's trigger when a user either opens a chat, or recieves a new message in the active chat.

It sends the conversation and message data needed to update the
`Conversation` model and broadcast to other clients.

When a client recieves an event, the home component updates the
conversation messages by updating what message gets the recipient
avatar.

API Fix
The API now correctly selects which message gets `userLastRead`
property.